画像追加

+++ face/views.py
@@ -6,25 +6,16 @@ from datetime import datetime
 import cv2
 import numpy as np
 from django.conf import settings
-from django.contrib.auth.decorators import login_required
 from django.http import HttpResponse
 from django.shortcuts import redirect, render
 from PIL import Image, ImageDraw, ImageFilter
 
 from .forms import ImageForm
-from .models import UploadedImage
-
-# def home(request):
-#     params = {}
-
-#     return render(request, "home.html", params)
 
 
 def index(request):
     params = {}
     params["form"] = ImageForm()
-    # params["uploaded_images"] = UploadedImage.objects.filter(author=request.user).order_by(
-    #     '-created_at')
 
     return render(request, "index.html", params)
 
@@ -33,30 +24,16 @@ def upload_image(request):
     if request.method == "POST":
         form = ImageForm(request.POST, request.FILES)
         if form.is_valid():
-            # uploaded_image = UploadedImage()
-            # uploaded_image.user_im = request.FILES.get("user_im")
-            # uploaded_image.product_im = request.FILES.get("product_im")
-            # uploaded_image.author = request.user
-            # uploaded_image.save()
             user_im = request.FILES.get("user_im")
             product_im = request.FILES.get("product_im")
 
             day, output = recognize_face(user_im, product_im)
-            # uploaded_image.image = get_image_path(uploaded_image)
-            # uploaded_image.save()
-            # path = str(settings.BASE_DIR) + "/media" + \
-            #     get_image_path()
+            if output == None:
+                return redirect('index')
+
             result_file = open(output, 'rb').read()
             delete_file(day)
     return HttpResponse(result_file, content_type="image/png")
-
-
-# def get_image_path(day):
-#     return "/face_images/%s%s" % (
-#         hashlib.sha1(
-#             (day).encode("utf-8")).hexdigest(),
-#         ".png",
-#     )
 
 
 def get_tmp_image_path(dir, day):
@@ -72,14 +49,12 @@ def get_tmp_image_path(dir, day):
 def recognize_face(user_im, product_im):
     day = str(datetime.utcnow().date()).replace('-', '')
     face_cascade = cv2.CascadeClassifier('opencv/face_cascade.xml')
-
-    # url = uploaded_image.user_im.url
-    # path = str(settings.BASE_DIR) + url
-    # src = cv2.imread(path)
     src = cv2.imdecode(np.fromstring(
         user_im.read(), np.uint8), cv2.IMREAD_UNCHANGED)
 
     faces = face_cascade.detectMultiScale(src)
+    if not faces.any():
+        return None, None
     face = sorted(faces, reverse=True, key=lambda x: x[2])[0]
 
     x, y, w, h = face[0], face[1], face[2], face[3]
@@ -105,11 +80,12 @@ def recognize_face(user_im, product_im):
     im_rgba_crop.save(crop_path)
 
     # 商品画像の顔を識別
-    # pro_path = str(settings.BASE_DIR) + uploaded_image.product_im.url
-    # pro_src = cv2.imread(pro_path)
     pro_src = cv2.imdecode(np.fromstring(
         product_im.read(), np.uint8), cv2.IMREAD_UNCHANGED)
     pro_faces = face_cascade.detectMultiScale(pro_src)
+
+    if not pro_faces.any():
+        return None, None
     pro_face = sorted(pro_faces, reverse=True, key=lambda x: x[2])[0]
 
     px, py, pw, ph = pro_face[0], pro_face[1], pro_face[2], pro_face[3]
@@ -117,9 +93,8 @@ def recognize_face(user_im, product_im):
     pro_im = cv2.cvtColor(pro_src, cv2.COLOR_BGR2RGB)
     pro_im = Image.fromarray(pro_im)
 
-    # pro_im = Image.open(pro_path)
     copy_pro_im = pro_im.copy()
-    im_rgba_crop = im_rgba_crop.resize((pw+sab*2, ph+sab*2))
+    im_rgba_crop = im_rgba_crop.resize((int(pw+sab*1.5), int(ph+sab*1.5)))
     copy_pro_im.paste(im_rgba_crop, (px-sab, py-sab),
                       im_rgba_crop.split()[3])
     copy_pro_im.save(str(settings.BASE_DIR) +
